move pagination into helper dir, add tests
