+++ src/types/pagination.ts
@@ -1,0 +1,16 @@
+export interface PageItem {
+    number?: number;
+    href?: string;
+    current?: boolean;
+    ellipsis?: boolean;
+  }
+
+export interface PaginationPreviousNext {
+    href: string;
+  }
+
+export interface PaginationData {
+    previous?: PaginationPreviousNext;
+    next?: PaginationPreviousNext;
+    items: PageItem[];
+  }
